requirements.txt
keras==2.11.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
scikit_learn==1.1.3
xgboost==1.6.2
